them search end page

+++ admin/product/list.php
@@ -1,20 +1,43 @@
 <?php
-$dir = str_replace("admin\product", "", __DIR__);  // hàm str_replace là hàm tha đổi giá trị đoạn chuỗi, mhư ở đây cắt phần admin/users
+$dir = str_replace("admin\product", "", __DIR__);  // hàm str_replace là hàm tha đổi giá trị đoạn chuỗi, mhư ở đây cắt phần admin/products
 require_once $dir . 'dals/productDAL.php';
 require_once $dir . 'config.php';
+$productDAL = new productDAL();
+   $resultNum = $productDAL->getList();
+    // trả về số sản phẩm có trong giỏ hàng
+$number = mysqli_num_rows($resultNum);
+//  hàm ceil là làm trò lên 
+$sotrang = ceil($number / 10);
+if (isset($_POST['min']) && $_POST['min'] != null) {
+    $min = $_POST['min'];
+    $max = $_POST['max'];
+    // hàm tìm sản phẩm trong column price
+    $result = $productDAL->getSearch($min, $max);
+} else {
+    if (isset($_GET['id']) && $_GET['id'] !== null) {
+        $id = $_GET['id'];
+        // truyên id sang để lấy vè số bản ghi tương ứng
+        $result = $productDAL->paging($id);
+    }else{
+        $id = 1;
+        $result = $productDAL->paging($id);
+    }
+}
 
-$userDAL = new productDAL();
-$result = $userDAL->getList();
+
 if (isset($_GET['action'])) {
     // hàm xóa
     if (is_numeric($_GET['id']) && $_GET['action'] == 'delete') {
         $id = $_GET['id'];
         $img = $_GET['img'];
-        $userDAL->deleteOne($id);
+        $productDAL->deleteOne($id);
         // lấy đường link của ảnh 
         $checked = $dir . $img;
         // hàm xóa đường link 
-        unlink($checked);
+        if ($checked) {
+            unlink($checked);
+        }
+
         header('location:list.php');
     }
 }
@@ -32,12 +55,12 @@ if (isset($_GET['action'])) {
         <?php require_once $dir . 'admin/commons/nav.php' ?>
         <div class="row">
             <!-- Left col -->
-            <section class="col-lg-7 connectedSortable">
+            <section class="col-lg-12 connectedSortable">
                 <!-- Custom tabs (Charts with tabs)-->
-                <div class="card">
+                <div class="card content1">
                     <div class="card-header">
                         <h3 class="card-title">
-                            <i class="fas fa-chart-pie mr-1"></i> User
+                            <i class="fas fa-chart-pie mr-1"></i> product
                         </h3>
 
                     </div>
@@ -45,157 +68,83 @@ if (isset($_GET['action'])) {
                     <div class="card-body">
                         <div class="tab-content p-0">
                             <!-- content -->
-                            <table border="1px" cellspacing=0 cellpadding=10>
-                        <thead>
-                            <tr>
-                                <th> id</th>
-                                <th>name</th>
-                                <th> price</th>
-                                <th>image</th>
-                                <th>content</th>
-                                <th>category_id</th>
-                            </tr>
-                        </thead>
-                        <?php
-                        foreach ($result as $row) :
-                        ?>
-                            <tbody>
-                                <tr>
-                                    <td>
-                                        <?php echo $row['product_id']; ?>
-                                    </td>
-                                    <td>
-                                        <?php echo $row['product_name']; ?>
-                                    </td>
-                                    <td>
-                                        <?php echo $row['price']; ?>
-                                    </td>
-                                    <td>
-                                        <img src="<?php echo BASE_URL . $row['image']; ?>" alt="" width="100">
+   
+                            <form action="" method="post" class="search">
+                                <label for="">min</label>
+                                <input type="text" name="min" placeholder="nhap vao gia tim kiem">
+                                <label for="">max</label>
+                                <input type="text" name="max" placeholder="nhap vao gia tim kiem">
 
-                                    </td>
-                                    <td>
-                                        <?php echo $row['content']; ?>
-                                    </td>
-                                    <td>
-                                        <?php echo $row['category_name']; ?>
-                                    </td>
-                                    <td>
-                                        <a href="edit.php?id=<?php echo $row['product_id']; ?>" class="btn btn-primary">sửa</a>
-                                        <a class="btn btn-danger" href="?action=delete&id=<?php echo $row['product_id'] ?>&img=<?php echo $row['image']; ?>">xóa</a>
-                                    </td>
-                                </tr>
-                            <?php endforeach; ?>
-                            <tr>
-                                <td>
-                                    <a class="btn btn-dark" href="add.php">thêm</a>
-                                </td>
-                            </tr>
-                            </tbody>
-                    </table>
+                                <button class="btn btn-dark">tim kiem</button>
+
+                            </form>
+                            <table border="1px" cellspacing=0 cellpadding=10>
+                                <thead>
+                                    <tr>
+                                        <th> id</th>
+                                        <th>name</th>
+                                        <th> price</th>
+                                        <th>image</th>
+                                        <th>content</th>
+                                        <th>category_id</th>
+                                    </tr>
+                                </thead>
+                                <?php
+                                foreach ($result as $row) :
+                                ?>
+                                    <tbody>
+                                        <tr>
+                                            <td>
+                                                <?php echo $row['product_id']; ?>
+                                            </td>
+                                            <td>
+                                                <?php echo $row['product_name']; ?>
+                                            </td>
+                                            <td>
+                                                <?php echo $row['price']; ?>
+                                            </td>
+                                            <td>
+                                                <img src="<?php echo BASE_URL . $row['image']; ?>" alt="" width="100">
+
+                                            </td>
+                                            <td>
+                                                <?php echo $row['content']; ?>
+                                            </td>
+                                            <td>
+                                                <?php echo $row['category_name']; ?>
+                                            </td>
+                                            <td>
+                                                <a href="edit.php?id=<?php echo $row['product_id']; ?>" class="btn btn-primary">sửa</a>
+                                                <a class="btn btn-danger" href="?action=delete&id=<?php echo $row['product_id'] ?>&img=<?php echo $row['image']; ?>">xóa</a>
+                                            </td>
+                                        </tr>
+                                    <?php endforeach; ?>
+                                    <tr>
+                                        <td>
+                                            <a class="btn btn-dark" href="add.php">thêm</a>
+                                        </td>
+                                    </tr>
+                                    </tbody>
+                            </table>  <?php
+                                        $i = 1;
+                                        while ($i <= $sotrang) { ?>
+                                        <div style="width: 100px; list-style:none; display:inline-block">
+                                             <li class="page-item <?php if (!$id) {
+                                                            $id = 1;
+                                                        }
+                                                        echo ($id == $i) ? 'active' : ''; ?>"><a style="text-align: center;" class="page-link" href="<?php $dir . 'admin/product/list.php' ?>?id=<?php echo $i ?>"><?php echo $i ?></a></li>
+                                        </div>
+                               
+                            <?php
+                                            $i++;
+                                        }
+                            ?>
                         </div>
                     </div>
                     <!-- /.card-body -->
                 </div>
                 <!-- /.card -->
             </section>
-            <!-- /.Left col -->
-            <!-- right col (We are only adding the ID to make the widgets sortable)-->
-            <section class="col-lg-5 connectedSortable">
-
-                <!-- Map card -->
-                <div class="card bg-gradient-primary">
-                    <div class="card-header border-0">
-                        <h3 class="card-title">
-                            <i class="fas fa-map-marker-alt mr-1"></i> Visitors
-                        </h3>
-                        <!-- card tools -->
-                        <div class="card-tools">
-                            <button type="button" class="btn btn-primary btn-sm daterange" title="Date range">
-                                <i class="far fa-calendar-alt"></i>
-                            </button>
-                            <button type="button" class="btn btn-primary btn-sm" data-card-widget="collapse" title="Collapse">
-                                <i class="fas fa-minus"></i>
-                            </button>
-                        </div>
-                        <!-- /.card-tools -->
-                    </div>
-                    <div class="card-body">
-                        <div id="world-map" style="height: 250px; width: 100%;"></div>
-                    </div>
-                    <!-- /.card-body-->
-                    <div class="card-footer bg-transparent">
-                        <div class="row">
-                            <div class="col-4 text-center">
-                                <div id="sparkline-1"></div>
-                                <div class="text-white">Visitors</div>
-                            </div>
-                            <!-- ./col -->
-                            <div class="col-4 text-center">
-                                <div id="sparkline-2"></div>
-                                <div class="text-white">Online</div>
-                            </div>
-                            <!-- ./col -->
-                            <div class="col-4 text-center">
-                                <div id="sparkline-3"></div>
-                                <div class="text-white">Sales</div>
-                            </div>
-                            <!-- ./col -->
-                        </div>
-                        <!-- /.row -->
-                    </div>
-                </div>
-                <!-- /.card -->
-
-                <!-- solid sales graph -->
-                <div class="card bg-gradient-info">
-                    <div class="card-header border-0">
-                        <h3 class="card-title">
-                            <i class="fas fa-th mr-1"></i> Sales Graph
-                        </h3>
-
-                        <div class="card-tools">
-                            <button type="button" class="btn bg-info btn-sm" data-card-widget="collapse">
-                                <i class="fas fa-minus"></i>
-                            </button>
-                            <button type="button" class="btn bg-info btn-sm" data-card-widget="remove">
-                                <i class="fas fa-times"></i>
-                            </button>
-                        </div>
-                    </div>
-                    <div class="card-body">
-                        <canvas class="chart" id="line-chart" style="min-height: 250px; height: 250px; max-height: 250px; max-width: 100%;"></canvas>
-                    </div>
-                    <!-- /.card-body -->
-                    <div class="card-footer bg-transparent">
-                        <div class="row">
-                            <div class="col-4 text-center">
-                                <input type="text" class="knob" data-readonly="true" value="20" data-width="60" data-height="60" data-fgColor="#39CCCC">
-
-                                <div class="text-white">Mail-Orders</div>
-                            </div>
-                            <!-- ./col -->
-                            <div class="col-4 text-center">
-                                <input type="text" class="knob" data-readonly="true" value="50" data-width="60" data-height="60" data-fgColor="#39CCCC">
-
-                                <div class="text-white">Online</div>
-                            </div>
-                            <!-- ./col -->
-                            <div class="col-4 text-center">
-                                <input type="text" class="knob" data-readonly="true" value="30" data-width="60" data-height="60" data-fgColor="#39CCCC">
-
-                                <div class="text-white">In-Store</div>
-                            </div>
-                            <!-- ./col -->
-                        </div>
-                        <!-- /.row -->
-                    </div>
-                    <!-- /.card-footer -->
-                </div>
-                <!-- /.card -->
-
-
-
         </div>
 
     </div>
